+++ src/component/titleLogo.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+
+const TitleLogo = () => {
+    return(
+      <h1 className="header">
+        <span>a</span>
+        <span>n</span>
+        <span>i</span>
+        <span>l</span>
+        <span>o</span>
+        <span>g</span>
+      </h1>
+    )
+  }
+
+  export default TitleLogo
